review fixes: split InterceptTab, extract observer helpers, AbortController drag, bounds comment
